Sync

+++ test/offline/db_engine_unit.js
@@ -15,34 +15,51 @@
  * limitations under the License.
  */
 
-describe('DBEngine', function() {
-  var db;
-  var schema;
-  var oldName;
+describe('DBEngine', /** @suppress {accessControls} */ function() {
+  /** @const */
+  var oldName = shaka.offline.DBEngine.DB_NAME_;
 
-  beforeAll(/** @suppress {accessControls} */ function() {
-    oldName = shaka.offline.DBEngine.DB_NAME_;
-    shaka.offline.DBEngine.DB_NAME_ += '_test';
+  /** @type {!shaka.offline.DBEngine} */
+  var db;
+  /** @type {!Object.<string, string>} */
+  var schema;
+
+  beforeAll(function() {
+    if (shaka.offline.DBEngine.isSupported()) {
+      shaka.offline.DBEngine.DB_NAME_ += '_test';
+    }
   });
 
   beforeEach(function(done) {
-    schema = {'test': 'key', 'other': 'key'};
-
-    shaka.offline.DBEngine.deleteDatabase().then(function() {
-      db = new shaka.offline.DBEngine();
-      return db.init(schema);
-    }).catch(fail).then(done);
+    if (shaka.offline.DBEngine.isSupported()) {
+      schema = {'test': 'key', 'other': 'key'};
+      shaka.offline.DBEngine.deleteDatabase().then(function() {
+        db = new shaka.offline.DBEngine();
+        return db.init(schema, /* opt_retryCount */ 5);
+      }).catch(fail).then(done);
+    } else {
+      done();
+    }
   });
 
-  afterAll(/** @suppress {accessControls} */ function() {
-    shaka.offline.DBEngine.DB_NAME_ = oldName;
+  afterAll(function() {
+    if (shaka.offline.DBEngine.isSupported()) {
+      shaka.offline.DBEngine.DB_NAME_ = oldName;
+    }
   });
 
   afterEach(function(done) {
-    db.destroy().catch(fail).then(done);
+    if (shaka.offline.DBEngine.isSupported()) {
+      db.destroy().catch(fail).then(done);
+    } else {
+      done();
+    }
   });
 
   it('stores and retrieves values', function(done) {
+    if (!shaka.offline.DBEngine.isSupported()) {
+      pending('DBEngine is not supported on this platform.');
+    }
     var data = {
       key: 123,
       extra: 'foobar'
@@ -55,6 +72,9 @@ describe('DBEngine', function() {
   });
 
   it('supports concurrent operations', function(done) {
+    if (!shaka.offline.DBEngine.isSupported()) {
+      pending('DBEngine is not supported on this platform.');
+    }
     var data1 = {key: 1, extra: 'cat'};
     var data2 = {key: 2, foobar: 'baz'};
     var data3 = {key: 3, abc: 123};
@@ -80,6 +100,9 @@ describe('DBEngine', function() {
   });
 
   it('supports remove', function(done) {
+    if (!shaka.offline.DBEngine.isSupported()) {
+      pending('DBEngine is not supported on this platform.');
+    }
     Promise.all([
       db.insert('test', {key: 1, i: 4}),
       db.insert('test', {key: 2, i: 1}),
@@ -93,7 +116,7 @@ describe('DBEngine', function() {
       return db.get('test', 2);
     }).then(function(data) {
       expect(data).toBeFalsy();
-      return db.removeWhere('test', function(s) { return s.i >= 7; });
+      return db.removeKeys('test', [4, 5, 6]);
     }).then(function() {
       return db.get('test', 5);
     }).then(function(data) {
@@ -104,8 +127,36 @@ describe('DBEngine', function() {
     }).catch(fail).then(done);
   });
 
+  it('supports iterating over each element', function(done) {
+    if (!shaka.offline.DBEngine.isSupported()) {
+      pending('DBEngine is not supported on this platform.');
+    }
+    var testData = [
+      {key: 1, i: 4},
+      {key: 2, i: 1},
+      {key: 3, i: 2},
+      {key: 4, i: 9}
+    ];
+    var spy = jasmine.createSpy('forEach');
+    Promise.all(testData.map(db.insert.bind(db, 'test')))
+        .then(function() {
+          return db.forEach('test', shaka.test.Util.spyFunc(spy));
+        })
+        .then(function() {
+          expect(spy).toHaveBeenCalledTimes(testData.length);
+          for (var i = 0; i < testData.length; i++)
+            expect(spy).toHaveBeenCalledWith(testData[i]);
+        })
+        .catch(fail)
+        .then(done);
+  });
+
   it('aborts transactions on destroy()', function(done) {
+    if (!shaka.offline.DBEngine.isSupported()) {
+      pending('DBEngine is not supported on this platform.');
+    }
     var expectedError = new shaka.util.Error(
+        shaka.util.Error.Severity.CRITICAL,
         shaka.util.Error.Category.STORAGE,
         shaka.util.Error.Code.OPERATION_ABORTED);
     var insert1Finished = false, insert2Finished = false;
@@ -133,6 +184,9 @@ describe('DBEngine', function() {
   });
 
   it('will find and reserve IDs', function(done) {
+    if (!shaka.offline.DBEngine.isSupported()) {
+      pending('DBEngine is not supported on this platform.');
+    }
     Promise
         .all([
           db.insert('test', {key: 1}),
@@ -159,5 +213,34 @@ describe('DBEngine', function() {
         })
         .catch(fail)
         .then(done);
+  });
+
+  it('will catch aborting transactions', function(done) {
+    if (!shaka.offline.DBEngine.isSupported()) {
+      pending('DBEngine is not supported on this platform.');
+    }
+
+    // Change the insert function so that once the put request completes
+    // the transaction will abort. This should cause the promise to be
+    // rejected.
+    db.insert = function(storeName, value) {
+      return db.createTransaction_(storeName, 'readwrite', function(store) {
+        var request = store.put(value);
+        request.onsuccess = function(event) {
+          request.transaction.abort();
+        };
+      });
+    };
+
+    var expected = new shaka.util.Error(
+        shaka.util.Error.Severity.CRITICAL,
+        shaka.util.Error.Category.STORAGE,
+        shaka.util.Error.Code.OPERATION_ABORTED);
+
+    db.insert('test', {key: 1}).then(fail, function(error) {
+      shaka.log.info('insert failed as expected ', error);
+      shaka.test.Util.expectToEqualError(error, expected);
+      done();
+    });
   });
 });
